Ported over the CogsLoading commands to a cog
